backend: shared types
