+++ backend/src/util/authentication.ts
@@ -2,7 +2,8 @@ import { Request } from 'express';
 import jwt, { JsonWebTokenError } from 'jsonwebtoken';
 import { JWT_SECRET } from './config';
 import { error, isNumber, ok } from './utils';
-import { Err, Ok, UserWithToken } from './types';
+import { Err, Ok } from './types';
+import { UserWithToken } from '../types/shared';
 import { UserModel } from '../models';
 
 type TokenMissing = { k: 'token-missing' };
